admin panel

+++ backend/src/main/java/ru/ispo/music_service/repository/TrackRepository.java
@@ -1,6 +1,5 @@
 package ru.ispo.music_service.repository;
 
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.ispo.music_service.entity.Track;
 import org.springframework.data.jpa.repository.Query;
@@ -10,7 +9,7 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
-public interface TrackRepository extends JpaRepository<Track, Long> {
+public interface TrackRepository extends JpaRepository<Track, Integer> {
     Optional<Track> findByTrackId(Integer trackId);
     @Query("SELECT t FROM Track t " +
             "JOIN Pricing p ON t.trackId = p.track.trackId " +
